Implement pitchBendRange conversion.
