+++ src/mpeInstrument/index.js
@@ -1,7 +1,7 @@
 import { createStore, applyMiddleware, compose } from 'redux';
 import { generateMidiActions, clearActiveNotes } from './actions';
 import { logger } from './middlewares';
-import { normalizeAction, addPitch } from './utils/actionUtils';
+import { normalize, addPitch, convertPitchBendRange } from './utils/activeNoteUtils';
 import rootReducer from './reducers';
 
 /**
@@ -40,11 +40,13 @@ export const mpeInstrument = options => {
     log: false,
     normalize: true,
     pitch: false,
+    pitchBendRange: 48,
   };
   const defaultedOptions = Object.assign({}, defaults, options);
   const formatNote = compose(...[
-    defaultedOptions.normalize && normalizeAction,
-    defaultedOptions.pitch && addPitch(options.pitch),
+    defaultedOptions.pitch && addPitch(defaultedOptions),
+    defaultedOptions.pitchBendRange && convertPitchBendRange(defaultedOptions),
+    defaultedOptions.normalize && normalize,
   ].filter(f => f));
   const formatActiveNotes = notes => notes.map(formatNote);
   const middlewares = [
